requirement.txt
cupy-cuda12x
Pillow